pdf_utilities/requirements.txt
camelot-py==0.9.0
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cryptography==41.0.4
cssselect==1.2.0
cycler==0.12.0
distro==1.8.0
et-xmlfile==1.1.0
fonttools==4.43.1
ghostscript==0.7
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.0
numpy==1.26.0
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.0.1
pycparser==2.21
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
pyparsing==3.1.1
PyPDF2==2.12.1
pyquery==2.0.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
